Hack: increase root USB hub reset time to reset Cam Link 4K 20GAM9902
20GAM9902 seems to turn off its USB interface unless enumerated
within a couple of seconds of being powered on. This causes the
20GAM9902 to not enumerate at boot time on a number of boards.

On Seeed Studio J1010, holding the RESET state of the root USB hub
power gates the USB ports, resetting the 20GAM9902 and allowing it
to be detected once the RESET is removed. A delay of at least 500ms
seems to be required for the 20GAM9902 to reset, but it's set here
to 1000ms to have a margin in case there's additional capacitance
on the bus