Hack: xhci-tegra: delay turning on VBUS to reset Cam Link 4K 20GAM9902
20GAM9902 seems to turn off its USB interface unless enumerated
within a couple of seconds of being powered on. This causes the
20GAM9902 to not enumerate at boot time on a number of boards.

On the Jetson Nano 2GB devkit, the USB ports are power gated
via a GPIO pin. We delay turning it on to allow enough time for
the 20GAM9902 to reset and for the the USB stack to initialize